backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
pydantic[email]==2.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.25.2
python-dotenv==1.0.0
redis==5.0.1
celery==5.3.4
email-validator==2.1.0
Pillow==10.0.0
